==> 4_Map_Algebra.R <==
############################################
# Please make reference of this work as: Gabriel E. García-Peña, André V. Rubio, Hugo Mendoza, Miguel Fernández, Matthew T. Milholland, A. Alonso Aguirre, Gerardo Suzán, Carlos Zambrana-Torrelio. 2021. Land-use change and rodent-borne diseases: Hazards on the shared socioeconomic pathways. Phil. Trans. R. Soc. B.
# Code developer: Gabriel E. García-Peña
#
#
# OBJECTIVES OF THIS R CODE:
# 1) Read zoonotic hazard for each of the land use scenarios, generated by the file 3_zoonotic_hazard_by_cell.R
# 2) Perform map algebra to calculate differences in hazard between scenarios 
# 3) Publish results in a raster
# 
################################################################
# Requirements:
# Files of zoonotic hazard generated by code in 3_zoonotic_hazard_by_cell.R

# HAZARD ON THE SUSTAINABLE PATH IN THRE YEARS: 2025,2050,AND 2100
cell.sum.SSP1.2025 = read.csv("cell_sum_N10_2025.csv") # SSP1: Sustainable development 2025
cell.sum.SSP1.2050 = read.csv("cell_sum_N10_2050.csv") # SSP1: Sustainable development 2050

# HAZARD ON THE SHARED SOCIOECONOMIC PATHWAYS, FOR THRE YEAR 2050
cell.sum.SSP4.2050 = read.csv("cell_sum_N10_SSP4_2050.csv") # SSP4: Inequality
cell.sum.SSP5.2050 = read.csv("cell_sum_N10_SSP5_2050.csv") # SSP5: fossil-fuelled development

# DATA ON RODENT RICHNES AT EACH GEOGRAPHIC CELL
richness = cell.sum.SSP4.2050$reservoir.richness
names(richness) = cell.sum.SSP4.2050$cell


# Assign variables as vectors named by the cell name
    SSP1.2025 <- cell.sum.SSP1.2025$cell.sum.reservoirs # BASELINE
    SSP1.2050 <- cell.sum.SSP1.2050$cell.sum.reservoirs
    SSP3.2050 <- cell.sum.SSP3.2050$cell.sum.reservoirs
    SSP4.2050 <- cell.sum.SSP4.2050$cell.sum.reservoirs
    SSP5.2050 <- cell.sum.SSP5.2050$cell.sum.reservoirs

# Assign names to the variables
    names(SSP1.2025) <- cell.sum.SSP1.2025$cell
    names(SSP1.2050) <- cell.sum.SSP1.2050$cell
    names(SSP3.2050) <- cell.sum.SSP3.2050$cell
    names(SSP4.2050) <- cell.sum.SSP4.2050$cell
    names(SSP5.2050) <- cell.sum.SSP5.2050$cell

# Generate an index with all cells with data
    idx = unique(c(names(SSP1.2025)
                   , names(SSP1.2050)
                   , names(SSP3.2050)
                   , names(SSP4.2050)
                   , names(SSP5.2050)
                   , names(SSP1.2100)
                   , names(SSP3.2100)
                   , names(SSP4.2100)
                   , names(SSP5.2100)
                   ))


# INDEX SHARED SOCIECONOMIC PATHWAYS (SSP) FOR MAP ALGEBRA

    SSP1_2025 = SSP1.2025[idx]
    names(SSP1_2025) = idx

    SSP1_2050 = SSP1.2050[idx]
    names(SSP1_2050) = idx

    SSP4_2050 = SSP4.2050[idx]
    names(SSP4_2050) = idx

    SSP5_2050 = SSP5.2050[idx]
    names(SSP5_2050) = idx


# HAZARD CHANGE UNDER CONSTRASTING SCENARIOS OF LAND USE
# SUSTAINABLE
    SSP1_2050_SSP1_2025 = SSP1.2050[idx]-SSP1.2025[idx]
    names(SSP1_2050_SSP1_2025) = idx

    # FOSSIL FUELLED SHARED SOCIOECONOMIC PATHWAY
    SSP5_2050_SSP1_2025 = SSP5.2050[idx]-SSP1.2025[idx]
    names(SSP5_2050_SSP1_2025) = idx

# HAZARD CHANGE IN A SSP4 OF DEEPENING INEQUALITY
# SPP5.2050 - SSP1.2050
    SSP4_2050_SSP1_2025 = SSP4.2050[idx]-SSP1.2025[idx]
    names(SSP4_2050_SSP1_2025) = idx

# AVOIDED COSTS
# SPP5.2050-SSP1.2050
    SSP5_2050_SSP1_2050 = SSP5.2050[idx]-SSP1.2050[idx]
    names(SSP5_2050_SSP1_2050) = idx

# SPP4.2050 - SSP1.2050
    SSP4_2050_SSP1_2050 = SSP4.2050[idx]-SSP1.2050[idx]
    names(SSP4_2050_SSP1_2050) = idx


# REPORT RESULTS IN RASTER
# CREATE RASTERS
# reference: https://stackoverflow.com/questions/19627344/how-to-create-a-raster-from-a-data-frame-in-r

library(raster)

# Create spatial points data frame
# CHOOSE THE DATA TO BE PUBLISHED IN THE RASTER. NOTE THAT THIS FUNCTION CAN BE PASSED TO A LIST. 

X = data.frame(richness) # HERE WE CAN CHOOSE THE VARIABLE TO PUBLISH IN A RASTER. FOR EXAMPLE richness

# Collect coordinates by decomposing the name of the cells
    coord = do.call(rbind, strsplit(rownames(X), " "))
    lat = as.numeric(coord[,1])
    lon = as.numeric(coord[,2])

    spg <- data.frame(X, lat, lon) 
    coordinates(spg) <- ~ lon + lat

# coerce to SpatialPixelsDataFrame
    gridded(spg) <- TRUE

# coerce to raster
    rasterDF <- raster(spg)
    plot(rasterDF)
    path_r= "PATH TO RESULTS"

# Format to save the data has to be specified, whether the data is integer, float, etc. 
# If data is integer: 
    writeRaster(rasterDF, file=paste(path_r, names(X), ".tif"), format = "GTiff", overwrite=T, datatype="INT1U")

# If data is between -50 to 50
    writeRaster(rasterDF, file=paste(path_r, names(X), ".tif"), format = "GTiff", overwrite=T, datatype="FLT4S")
